pyproject.toml
[tool.poetry]
name = "football-data"
version = "0.1.0"
description = "Project for scraping and processing football data"
authors = ["mateusz-jalocha <m.jalocha@numlabs.com>"]
readme = "README.md"
packages = [{include = "football_data"}]

[tool.poetry.dependencies]
python = "3.10.4"
beautifulsoup4 = "4.11.1"
bs4 = "0.0.1"
certifi = "2022.9.24"
charset-normalizer = "2.1.1"
idna = "3.4"
joblib = "1.2.0"
numpy = "1.23.3"
pandas = "1.5.0"
pydantic = "1.10.2"
python-dateutil = "2.8.2"
pytz = "2022.4"
requests = "2.28.1"
scikit-learn = "1.1.2"
scipy = "1.9.1"
six = "1.16.0"
sklearn = "0.0"
soupsieve = "2.3.2.post1"
threadpoolctl = "3.1.0"
typing-extensions = "4.4.0"
urllib3 = "1.26.12"
pytest = "^7.1.3"
jupyter = "^1.0.0"
jupytext = "^1.14.1"
bump2version = "^1.0.1"
python-dotenv = "^0.21.0"
logging = "^0.4.9.6"
pre-commit = "^2.20.0"
lxml = "^4.9.1"
isort = "^5.10.1"
black = "^22.10.0"

[tool.poetry.dev-dependencies]
bump2version = "*"
pre-commit = "^2.20.0"
pytest = "^7.1.3"
jupyter = "*"
jupytext = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
